Add files via upload
Added a button that should open priority queue, requires further coding for actual functionality.
