+++ Clock-master/Clock-master/nbproject/private/private.xml
@@ -4,10 +4,12 @@
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/View.java</file>
+            <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/PriorityItem.java</file>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/Clock.java</file>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/ListNode.java</file>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/menuBar.java</file>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/Model.java</file>
+            <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/Person.java</file>
             <file>file:/F:/BSc%20Computing(Software)%202018-2019/Software%20Construction/Assignment%202/Clock-master/Clock-master/src/clock/Controller.java</file>
         </group>
     </open-files>
